backend/requirements.txt
# Core
django>=5.1,<6.0
django-ninja>=1.3,<2.0

# MCP (Model Context Protocol) server
mcp>=1.0,<2.0
channels>=4.0,<5.0
daphne>=4.0,<5.0

# Socket.IO (Runner WebSocket protocol)
python-socketio>=5.0,<6.0

# CORS
django-cors-headers>=4.0,<5.0

# Static files
whitenoise>=6.0,<7.0

# Settings & Environment
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0

# JWT
PyJWT>=2.8,<3.0

# Encryption
cryptography>=43.0,<45.0

# Logging
structlog>=24.0

# Production database & caching
psycopg2-binary>=2.9,<3.0
channels-redis>=4.0,<5.0

# Dev & Testing
pytest>=8.0
pytest-django>=4.0
pytest-asyncio>=0.24
ruff>=0.8
